Added the clear filters button to the lower part of the component tree to make the process more efficeint
